+++ dialbb/builtin_blocks/chatgpt/chatgpt.py
@@ -25,14 +25,25 @@ __copyright__ = 'C4A Research Institute, Inc.'
 import os
 import sys
 import traceback
+import datetime
 from typing import Dict, Any, List, Union, Tuple
 import openai
 from dialbb.abstract_block import AbstractBlock
 from dialbb.util.error_handlers import abort_during_building
+import re
+
+from dialbb.util.globals import CHATGPT_INSTRUCTIONS
 
 DIALOGUE_HISTORY_OLD_TAG: str = '@dialogue_history'
 DIALOGUE_HISTORY_TAG: str = '{dialogue_history}'
+CURRENT_TIME_TAG: str = '{current_time}'
 DEFAULT_GPT_MODEL: str = "gpt-4o-mini"
+DIALOGUE_UP_TO_NOW = {"ja": "現在までの対話", "en": "Dialogue up to now"}
+
+
+#  [[[....{tag1}....{tag2}....]]]
+REMAINING_TAGS_PATTERN = re.compile( r"\[\[\[(?=.*\{[A-Za-z0-9_]+\})(?:[^\{\]]|\{[A-Za-z0-9_]+\})*\]\]\]",
+                                     re.DOTALL)
 
 
 class ChatGPT(AbstractBlock):
@@ -50,6 +61,12 @@ class ChatGPT(AbstractBlock):
         self._openai_client = openai.OpenAI(api_key=openai_api_key)
         self._gpt_model = self.block_config.get("gpt_model", DEFAULT_GPT_MODEL)
 
+        self._language = self.config.get("language", 'en')
+        self._instruction = self.block_config.get("instruction", CHATGPT_INSTRUCTIONS[self._language])
+
+        self.user_name: str = self.block_config.get("user_name", 'ユーザ' if self._language == 'ja' else "User")
+        self.system_name: str = self.block_config.get("system_name", 'システム' if self._language == 'ja' else "System")
+
         # reading prompt template file
         prompt_template_file: str = self.block_config.get("prompt_template", "")
         if not prompt_template_file:
@@ -64,6 +81,8 @@ class ChatGPT(AbstractBlock):
 
         # temperature
         self._temperature = self.block_config.get("temperature", 0.7)
+
+
 
         # {"session1" : [{"speaker": "user", "utterance": <user utterance>},
         #                {"speaker": "system", "utterance": <system utterance>},
@@ -130,6 +149,25 @@ class ChatGPT(AbstractBlock):
         system_utterance: str = chat_completion.choices[0].message.content
         return system_utterance
 
+    @staticmethod
+    def get_current_time_string(language: str) -> str:
+        """
+        generates string to represent current string.
+        :param language:
+        :return:
+        """
+
+        now = datetime.datetime.now()
+        if language == 'ja':
+            weekdays = ["月", "火", "水", "木", "金", "土", "日"]
+            date_str = now.strftime("%Y年%m月%d日")
+            time_str = now.strftime("%H時%M分%S秒")
+            weekday_str = weekdays[now.weekday()]
+            result: str = f"{date_str}（{weekday_str}） {time_str}"
+        else:
+            result = now.strftime("%A, %B %d, %Y %I:%M:%S %p")
+        return result
+
     def generate_system_utterance(self, dialogue_history: List[Dict[str, str]],
                                   session_id: str, user_id: str,
                                   aux_data: Dict[str, Any]) -> Tuple[str, Dict[str, Any], bool]:
@@ -144,18 +182,42 @@ class ChatGPT(AbstractBlock):
         :return: a tuple of system utterance string, aux_data as is, and final flag (always False)
         """
 
-        messages = []
-        messages.append({'role': "system", "content": self._prompt_template})
+        language = self.config.get("langauge", 'en')
+        prompt = self._prompt_template
+        prompt = prompt.replace(CURRENT_TIME_TAG, self.get_current_time_string(language))  # {current_time}
+        if aux_data:
+            for aux_data_key, aux_data_value in aux_data.items():  # aux_data values replace their place holders
+                prompt = prompt.replace("{" + aux_data_key + "}", str(aux_data_value))
+        prompt = REMAINING_TAGS_PATTERN.sub("", prompt)  # remove remaining tags enclosed by [[[ .... ]]]]
+        prompt = prompt.replace('[[[', "")  # remove remaining brackets
+        prompt = prompt.replace(']]]', "")
 
+
+        # add dialogue history to string
+        dialogue_history_string: str = ""
         for turn in dialogue_history:
             if turn["speaker"] == 'user':
-                messages.append({'role': "user", "content": turn['utterance']})
+                dialogue_history_string += f"{self.user_name}: {turn['utterance']}\n"
             else:
-                messages.append({'role': "assistant", "content": turn['utterance']})
+                dialogue_history_string += f"{self.system_name}: {turn['utterance']}\n"
 
+        if prompt.find(DIALOGUE_HISTORY_TAG) >= 0:
+            prompt: str = self._prompt_template.replace(DIALOGUE_HISTORY_TAG, dialogue_history_string)
+        elif prompt.find(DIALOGUE_HISTORY_OLD_TAG) >= 0:
+            prompt: str = prompt.replace(DIALOGUE_HISTORY_OLD_TAG, dialogue_history_string)
+        else:
+            prompt += f"\n#{DIALOGUE_UP_TO_NOW[language]}\n\n{dialogue_history_string}"
+
+        # create messages
+        messages = []
+        messages.append({'role': "system", "content": self._instruction})
+        messages.append({'role': "user", "content": prompt})
         self.log_debug("messages: " + str(messages), session_id=session_id)
-        system_utterance: str = self._generate_with_openai_gpt(messages)
-        self.log_debug("generated system utterance: " + system_utterance, session_id=session_id)
+
+        generated_utterance: str = self._generate_with_openai_gpt(messages)
+        self.log_debug("generated system utterance: " + generated_utterance, session_id=session_id)
+        system_utterance: str = generated_utterance.replace(f'{self.system_name}:', '').strip()
+        self.log_debug("final system utterance: " + system_utterance, session_id=session_id)
 
         return system_utterance, aux_data, False