+++ dialbb/builtin_blocks/stn_management/builtin_scenario_functions.py
@@ -23,12 +23,18 @@ __version__ = '0.1'
 __author__ = 'Mikio Nakano' 
 __copyright__ = 'C4A Research Institute, Inc.'
 
-import sys
 import traceback
 from typing import Dict, Any, List
 import os
 import datetime
+import re
+from dialbb.builtin_blocks.stn_management.util import scenario_function_log_debug
 
+from dialbb.util.globals import CHATGPT_INSTRUCTIONS
+
+#  [[[....{<tag1>}....{<tag2>}....]]]
+REMAINING_TAGS_PATTERN = re.compile( r"\[\[\[(?=.*\{[A-Za-z0-9_]+\})(?:[^\{\]]|\{[A-Za-z0-9_]+\})*\]\]\]",
+                                     re.DOTALL)
 
 DEFAULT_GPT_MODEL = 'gpt-4o-mini'
 
@@ -269,7 +275,8 @@ def call_chatgpt(prompt: str, context: Dict[str, Any], checking: bool = False) -
                                                           chatgpt_settings.get("temperature", 0.7))
         else:  # generation
             gpt_temperature: float = chatgpt_settings.get("temperature", 0.7)
-        instruction: str = chatgpt_settings.get("instruction", "")
+        language = context['_config'].get('language', 'en')
+        instruction: str = chatgpt_settings.get("instruction", CHATGPT_INSTRUCTIONS[language])
 
     else:
         gpt_model: str = DEFAULT_GPT_MODEL
@@ -284,14 +291,22 @@ def call_chatgpt(prompt: str, context: Dict[str, Any], checking: bool = False) -
         messages: List[Dict[str, str]] = [{"role": "user", "content": prompt}]
     while True:
         try:
-            temperature = 1 if gpt_model == 'gpt-5' else gpt_temperature
-            if DEBUG:
-                print(f"calling chatgpt: model: {gpt_model}, temperature: {str(temperature)}, messages: {str(messages)}")
-            chat_completion = openai_client.with_options(timeout=10).chat.completions.create(
-                model=gpt_model,
-                messages=messages,
-                temperature=temperature,
-            )
+            if gpt_model.startswith("gpt-4") or gpt_model.startswith("gpt-3"):
+                scenario_function_log_debug(f"calling chatgpt: model: {gpt_model}, " +
+                                            f"temperature: {str(gpt_temperature)}, messages: {str(messages)}",
+                                            context)
+                chat_completion = openai_client.with_options(timeout=10).chat.completions.create(
+                    model=gpt_model,
+                    messages=messages,
+                    temperature=gpt_temperature,
+                )
+            else:
+                scenario_function_log_debug(f"calling chatgpt: model: {gpt_model}, messages: {str(messages)}",
+                                            context)
+                chat_completion = openai_client.with_options(timeout=10).chat.completions.create(
+                    model=gpt_model,
+                    messages=messages
+                )
         except openai.APITimeoutError:
             continue
         except Exception as e:
@@ -307,29 +322,31 @@ def call_chatgpt(prompt: str, context: Dict[str, Any], checking: bool = False) -
 
 
 def get_current_time_string(language: str) -> str:
-    
+
     now = datetime.datetime.now()
     if language == 'ja':
         weekdays = ["月", "火", "水", "木", "金", "土", "日"]
         date_str = now.strftime("%Y年%m月%d日")
         time_str = now.strftime("%H時%M分%S秒")
         weekday_str = weekdays[now.weekday()]
+        result: str = f"{date_str}（{weekday_str}） {time_str}"
     else:
-        weekdays = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
-        date_str = now.strftime("%Y-%m-%d")
-        time_str = now.strftime("%H:%M:%S")
-        weekday_str = weekdays[now.weekday()]
-    result: str = f"{date_str}（{weekday_str}） {time_str}"
+        result = now.strftime("%A, %B %d, %Y %I:%M:%S %p")
     return result
 
 
-def create_prompt_from_template(prompt_template: str, dialogue_history_string: str,
-                                situation: str, persona: str, language: str) -> str:
+def create_prompt_from_template(prompt_template: str,
+                                situation: str, persona: str,
+                                dialogue_history_string: str,
+                                language: str, aux_data: Dict[str, Any]) -> str:
     """
     create prompt from prompt template
-    :param prompt_template:
-    :param dialogue_history_string:
+    :param prompt_template: prompt template
+    :param dialogue_history_string: stringified dialouge history
+    :param situation: stringified situation for chatgpt
+    :param persona: stringified persona for chatgpt
     :param language: 'ja', 'en', ...
+    :param aux_data: aux_data from main module
     :return: prompt
     """
 
@@ -342,6 +359,11 @@ def create_prompt_from_template(prompt_template: str, dialogue_history_string: s
     prompt = prompt.replace("@current_time@", get_current_time_string(language))
     prompt = prompt.replace("@persona@", persona)
     prompt = prompt.replace("@situation@", situation)
+    for aux_data_key, aux_data_value in aux_data.items():
+        prompt = prompt.replace("{" + aux_data_key + "}", str(aux_data_value))  # aux_data values replace their place holders
+    prompt = REMAINING_TAGS_PATTERN.sub("", prompt)  # remove remaining tags enclosed by [[[ .... ]]]
+    prompt = prompt.replace('[[[', "")  # remove remaining brackets
+    prompt = prompt.replace(']]]', "")
     return prompt
 
 
@@ -393,12 +415,15 @@ def builtin_generate_with_prompt_template(prompt_template: str, context: Dict[st
 
     dialogue_history: List[Dict[str, str]] = context['_dialogue_history']
     dialogue_history_string: str = create_dialogue_history_string(dialogue_history, language)
+    aux_data = context['_aux_data']
     situation, persona = get_situation_and_persona_from_config(context)
-    prompt: str = create_prompt_from_template(prompt_template, situation, persona, dialogue_history_string, language)
+    prompt: str = create_prompt_from_template(prompt_template, situation, persona,
+                                              dialogue_history_string, language, aux_data)
+
+    scenario_function_log_debug("prompt for chatgpt: " + prompt, context)
     generated_utterance = call_chatgpt(prompt, context, checking=False)
 
-    if DEBUG:
-        print("utterance generated by chatgpt: " + generated_utterance, flush=True)
+    scenario_function_log_debug("utterance generated by chatgpt: " + generated_utterance, context)
     generated_utterance = generated_utterance.replace(f'{system_name}:', '').strip()  # remove 'System: '
 
     return generated_utterance
@@ -435,12 +460,13 @@ def builtin_check_with_prompt_template(prompt_template: str, context: Dict[str, 
     language: str = context['_config'].get("language")
     dialogue_history: List[Dict[str, str]] = context['_dialogue_history']
     dialogue_history_string: str = create_dialogue_history_string(dialogue_history, language)
+    aux_data = context['_aux_data']
     situation, persona = get_situation_and_persona_from_config(context)
-    prompt: str = create_prompt_from_template(prompt_template, situation, persona, dialogue_history_string, language)
+    prompt: str = create_prompt_from_template(prompt_template, situation, persona,
+                                              dialogue_history_string, language, aux_data)
     response = call_chatgpt(prompt, context, checking=True)
 
-    if DEBUG:
-        print("check result: " + response, flush=True)
+    scenario_function_log_debug("check result: " + response, context)
     response = response.lower()
     if response.find("yes") == -1:  # yes was not found
         return False