+++ dialbb/builtin_blocks/stn_management/stn_manager.py
@@ -40,7 +40,7 @@ from dialbb.builtin_blocks.stn_management.context_db import ContextDB
 from dialbb.builtin_blocks.stn_management.state_transition_network \
     import StateTransitionNetwork, State, Transition, Argument, Condition, Action, \
     INITIAL_STATE_NAME, ERROR_STATE_NAME, FINAL_ABORT_STATE_NAME, GOSUB, EXIT, function_call_pattern, \
-    BUILTIN_FUNCTION_PREFIX, COMMA, SEMICOLON
+    BUILTIN_FUNCTION_PREFIX, COMMA, SEMICOLON, LEFTBRACE, RIGHTBRACE
 from dialbb.builtin_blocks.stn_management.stn_creator import create_stn
 from dialbb.abstract_block import AbstractBlock
 from dialbb.main import ANY_FLAG, DEBUG, CONFIG_KEY_FLAGS_TO_USE, CONFIG_DIR
@@ -80,6 +80,8 @@ CONTEXT_KEY_SUB_DIALOGUE_STACK: str = '_sub_dialogue_stack'
 CONTEXT_KEY_REACTION: str = '_reaction'
 CONTEXT_KEY_REQUESTING_CONFIRMATION: str = '_requesting_confirmation'
 CONTEXT_KEY_TURNS_IN_STATE: str = '_turns_in_state'
+CONTEXT_KEY_SESSION_ID: str = '_session_id'
+CONTEXT_KEY_USER_ID: str = '_user_id'
 
 INPUT_KEY_AUX_DATA: str = "aux_data"
 INPUT_KEY_SENTENCE: str = "sentence"
@@ -358,6 +360,8 @@ class Manager(AbstractBlock):
                            CONTEXT_KEY_SUB_DIALOGUE_STACK: [],
                            CONTEXT_KEY_REACTION: "",
                            CONTEXT_KEY_TURNS_IN_STATE: 1,
+                           CONTEXT_KEY_SESSION_ID: session_id,
+                           CONTEXT_KEY_USER_ID: user_id,
                            CONTEXT_KEY_REQUESTING_CONFIRMATION: False}
 
                 self._add_context(session_id, context)
@@ -480,6 +484,7 @@ class Manager(AbstractBlock):
                     self.log_debug("Skip state. Making another transition.", session_id=session_id)
                     new_state_name = self._transition(new_state_name, nlu_result, aux_data, context,
                                                       user_id, session_id, sentence)
+                    new_state_name = self._handle_sub_dialogue(new_state_name, session_id, context)
                     if not self._network.is_skip_state(new_state_name):
                         break
 
@@ -584,7 +589,7 @@ class Manager(AbstractBlock):
 
         for m in PROMPT_TEMPLATE_PATTERN.finditer(result):  # $$$ ... $$$
             prompt_template: str = m.group(1)
-            prompt_template = prompt_template.replace('{', '@').replace('}', '@')
+            prompt_template = prompt_template.replace('{', LEFTBRACE).replace('}', RIGHTBRACE)
             result = result.replace(m.group(0), f'{{_generate_with_prompt_template("{prompt_template}")}}')
 
         result = LLM_INSTRUCTION_PATTERN.sub(r'{_generate_with_llm("\1")}', result)   # $ ... $
@@ -794,7 +799,7 @@ class Manager(AbstractBlock):
                                           user_id, session_id, sentence)
                    for raw_argument in raw_argument_values]
         except Exception as e:  # failure in realization
-            self.log_warning(f"Exception occurred during realizing arguments in system utterance: {str(argument_names)}",
+            self.log_warning(f"Exception occurred during realizing arguments in system utterance: {str(raw_argument_values)}",
                              session_id=session_id)
             if DEBUG:
                 raise Exception(e)